requirements.txt
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
beanie==1.29.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
fastapi-mail==1.4.2
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongoengine==0.29.1
motor==3.7.0
passlib==1.7.4
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.13.2
shellingham==1.5.4
sniffio==1.3.1
starlette==0.45.3
toml==0.10.2
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
websockets==14.2
aiosmtplib==3.0.2
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.0.1
beanie==1.29.0
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
cssselect2==0.8.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
fastapi-mail==1.4.2
fonttools==4.56.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
googlesearch-python==1.3.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
lazy-model==0.2.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mongoengine==0.29.1
motor==3.7.0
passlib==1.7.4
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydyf==0.11.0
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11
pyparsing==3.2.1
PyPDF2==3.0.1
pyphen==0.17.2
python-dotenv==1.0.1
python-multipart==0.0.20
PyYAML==6.0.2
razorpay==1.4.2
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
rsa==4.9
setuptools==76.0.0
shellingham==1.5.4
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
weasyprint==64.1
webencodings==0.5.1
websockets==14.2
yt-dlp==2025.2.19
zopfli==0.2.3.post1
